+++ src/main/java/gift/config/RepositoryConfig.java
@@ -1,7 +1,7 @@
 package gift.config;
 
 import gift.model.repository.ProductRepository;
-import gift.model.Products;
+import gift.model.dao.MapProductDao;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -10,7 +10,7 @@ public class RepositoryConfig {
 
     @Bean
     public ProductRepository products(){
-        return new Products();
+        return new MapProductDao();
     }
 
 }
